Update tests for flexberry-lookup
