+++ tests/acceptance/components/flexberry-lookup/flexberry-lookup-projection-test.js
@@ -1,0 +1,35 @@
+import Ember from 'ember';
+import { executeTest } from './execute-flexberry-lookup-test';
+
+executeTest('flexberry-lookup projection test', (store, assert, app) => {
+  assert.expect(2);
+
+  visit('components-acceptance-tests/flexberry-lookup/settings-example-projection');
+
+  andThen(function() {
+    assert.equal(currentURL(), 'components-acceptance-tests/flexberry-lookup/settings-example-projection');
+
+    let $lookupButtouChoose = Ember.$('.ui-change');
+
+    // Click choose button.
+    Ember.run(() => {
+      $lookupButtouChoose.click();
+    });
+
+    Ember.run(() => {
+      var done = assert.async();
+      setTimeout(function() {
+
+        let $lookupSearch = Ember.$('.content table.object-list-view');
+        let $lookupSearchThead = $lookupSearch.children('thead');
+        let $lookupSearchTr = $lookupSearchThead.children('tr');
+        let $lookupHeaders = $lookupSearchTr.children('th');
+
+        // Check count at table header.
+        assert.strictEqual($lookupHeaders.length === 3, true, 'Component has SuggestionTypeE projection');
+
+        done();
+      }, 1000);
+    });
+  });
+});
